Read data from scripts incrementally and append to mutable data. I think the other method can block if the pipe gets filled up.
Only lock around the currentTask ivar, since the local task variable
is used for most things.  Run the runloop in its default mode until
the task is finished, instead of using waitUntilExit. We're not 
blocking the main thread, so the special runloop mode shouldn't be
needed.  

We also now respect the task's terminationStatus, and use
that as a signal for returning an error, and try defaultCStringEncoding
instead of ASCII if UTF-8 fails.